app-layer: support reset of application layer
Some protocols can change on the same connection. This for example
the case of HTTPS through a proxy (from HTTP to HTTPS). This is
also the case for SMTP via the STARTTLS command.
This patch allows an application layer to update the flow flag to
unset FLOW_ALPROTO_DETECT_DONE. If this is done, the protocol
recognition layer is redone to be able to detect the new protocol.