packet handling: fix release function
Extended data were freed before the release function was called.
The result was that, in AF_PACKET IPS mode, the release function
was only sending void data.

This patch updates the code to have the freeing of extended data
done in the cleaning function for a packet which is called by the
release function. This improve consistency of the code and fixes
the bug.