af-packet: add packet-size option
This option can be used to set the maximum datagram size that
af-packet can receive from an interface.

It could be useful to set it to a higher value than MTU is some offloading
is activated on the card and can not be deactivated.

Tœhis setup can also be used to be able to set default_packet_size
to 0 and avoid allocation when using zero copy.
+    #packet-size: 1522