app-layer-modbus: update client side tx if needed
In case the first data of a midstream picked Modbus session,
modbus code is adding a unsolicited event to the transaction.
Following that we see two successive unsollicited alerts for
the same flow.

This is due to the fact that the transaction inspected ID is
not updated because we have miss the client message.

To fix that we update the transaction id in the TO_SERVER direction
when midstream is enable and when we are at the beginning of the
session.

As in that case we should not alert, we also only set the
unsolicited event if we are not in this case.