af-packet: fix test on fanout availability
The fanout test was using the default 99 value for cluster_id.
This means that if another instance was running on cluster id 99
then it was not possible to run the current suricata with fanout.

So this patch updates the code to use the defined cluster_id as
parameter for the test. It also updates the message to signal there
is 2 possible solutions to the problem.