decode: introduce PACKET_ALERT_SRC_IS_TARGET flag
This patch converts to the packet alert flag instanciation to BIT_U8
and introduces PACKET_ALERT_SRC_IS_TARGET and PACKET_ALERT_DEST_IS_TARGET
that will be used to know which of the source or the destination IP is
the target of the attack. The flags will be used lateer to do some
postprocessing on generated alerts.